[SDPA-4021] Splitted tide page middleware js to different files
